Add idToHashid method to HasHashid trait
Converting an id to a corresponding hashid could be useful if you get an id from eg a foreign key. This function allows you to generate the hashid without first having to fetch the model from the database, or to first set an id on the model and then call ->hashid().

Just a bit more convenient.

Testing is covered in the test of $this->hashid() method, as that method is using it.